trie: move prune verifier as member into trie
